creates controllers for schema, table and all_db
